requirements.txt
opencv-python==4.8.1.78
tensorflow-macos==2.13.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
pillow==10.0.0
streamlit==1.25.0
plotly==5.15.0
python-dotenv==1.0.0
